requirements.txt
requests==2.31.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
mplfinance>=0.12.10b0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
